Update build scripts.

+++ scripts/unix/dev.sh
@@ -17,7 +17,11 @@ fi
 if [ "$TARGET" = "all" ];
 then
 
-    rm -rf dist;
+    if [ -d "dist" ];
+    then
+      rm -rf dist;
+    fi
+
     echo "Running webpack...";
     webpack --config webpack.config.js;
